A naive port to cats-effect 3
I couldn't find the equivalent of cats.effect.laws.util.TestContext, so instead
I made Env be the same as DefaultEnv for now (thus using real and not mocked
time for all tests).

Of course, this means that the tests run much longer, and the comparisons in
RateLimitingSpec have to be imprecise (and there is always a possibility of
failure).